pyproject.toml
[tool.poetry]
name = "thedailyglobetestdependency"
version = "0.1.0"
description = ""
authors = ["thelonggoodbuy <thelonggoodbuy1991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.5"
fastapi = "^0.112.2"
make = "^0.1.6.post2"
uvicorn = "^0.30.6"
dishka = "^1.3.0"
sqlalchemy = "^2.0.32"
advanced-alchemy = "^0.19.3"
alembic = "^1.13.2"
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
sqlalchemy-file = "^0.6.0"
mypy = "^1.11.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
authlib = "^1.3.2"
httpx = "^0.27.2"
itsdangerous = "^2.2.0"
litestar = "^2.11.0"
python-multipart = "^0.0.9"
email-validator = "^2.2.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
django-unfold = "^0.39.0"
psycopg2-binary = "^2.9.9"
django = "4.1"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
